add edit profile for user
I added tables to the database, added rows and
columns to all of the forms, and added a basic
panel.

I commented out the nested addresses for the user
because i'm not sure if we will have a need for
a polymorphic relationship(users and companies).

I also removed the call to googleapis for fonts.